requirements.txt
chainlit
litellm
google-generativeai
python-dotenv 